@@ -8,12 +8,12 @@
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.6.0",
     "format-duration": "^1.3.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+  "react": "^19.0.0",
+  "react-dom": "^19.0.0",
     "react-icons": "^4.1.0",
-    "react-redux": "^7.2.2",
+  "react-redux": "^8.1.1",
     "react-reveal": "^1.2.2",
-    "react-scripts": "4.0.1",
+  "react-scripts": "^5.0.1",
     "styled-components": "^5.2.1",
     "vtt-json": "^1.0.1",
     "web-vitals": "^0.2.4"
